server/requirements.txt
fastapi
uvicorn
pydantic
pytest
coverage
python-multipart
httpx
jinja2
black
colorama
google-generativeai
python-dotenv
